requirements.txt
altair==5.5.0
altgraph==0.17.4
asttokens==3.0.0
attrs==25.3.0
blinker==1.9.0
cachetools==6.1.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
decorator==5.2.1
exceptiongroup==1.3.0
executing==2.2.0
gitdb==4.0.12
GitPython==3.1.45
graphviz==0.21
idna==3.10
ipython==8.37.0
isodate==0.7.2
jedi==0.19.2
Jinja2==3.1.6
jsonpickle==4.1.1
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
narwhals==2.1.2
networkx==3.4.2
numpy==2.2.6
packaging==25.0
pandas==2.3.2
parso==0.8.5
pexpect==4.9.0
pillow==11.3.0
plotly==6.3.0
prompt_toolkit==3.0.51
protobuf==6.32.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==21.0.0
pydeck==0.9.1
pydot==4.0.1
Pygments==2.19.2
pyinstaller==6.15.0
pyinstaller-hooks-contrib==2025.8
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
pyvis==0.3.2
rdflib==7.1.4
referencing==0.36.2
requests==2.32.5
rpds-py==0.27.0
six==1.17.0
smmap==5.0.2
stack-data==0.6.3
streamlit==1.48.1
streamlit-agraph==0.0.45
tenacity==9.1.2
toml==0.10.2
tornado==6.5.2
traitlets==5.14.3
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
watchdog==6.0.0
wcwidth==0.2.13
